feat(base64): define export files with json object
Kubernetes secrets are key value pairs, where the filename becomes the
key and the value the file contents. In order to create an environment
file and make it easy for editing, the iondrive cli will interpret any
secret key starting with "env" and it's value being a json object as
configuration to create an environment file.